requirements.txt
pillow
tqdm
loguru
translate
faker

pandas

gdown
imutils

psutil
sklearn